@@ -59,8 +59,8 @@ class FractalChaosBandSeries extends Series {
       return ChannelFillPainter(
         fcbHighSeries,
         fcbLowSeries,
-        firstUpperChannelFillColor: config.fillColor.withOpacity(0.2),
-        secondUpperChannelFillColor: config.fillColor.withOpacity(0.2),
+        firstUpperChannelFillColor: config.fillColor.withValues(alpha: 0.2),
+        secondUpperChannelFillColor: config.fillColor.withValues(alpha: 0.2),
       );
     }
 
